--- a/src/csgoproto/Cargo.toml
+++ b/src/csgoproto/Cargo.toml
@@ -1,13 +1,14 @@
-
 [package]
 name = "csgoproto"
 version = "0.1.5"
 authors = ["Laiho"]
-
 edition = "2021"
+
+[lib]
+crate-type = ["lib"]
+
 [dependencies]
-bytes = "1.5.0"
-glob = "0.3.0"
-proc-macro2 = "1.0.78"
-protobuf = { version = "3.5.1", features = ["bytes"] }
-# protobuf-codegen = "3.5.0"
+prost = "0.13.3"
+
+[build-dependencies]
+prost-build = "0.13.3"