Seems to perform better than protobuf.

Python benchmark done on Ryzen 7 3700X.
12 demos (5.29Gb) parsed 10 times using `for i in range(10)`:
| Run # | Pre-commit time, in s | Post-commit time, in s | Difference |
|--------|--------|--------|--------|
| 1 | 6.94 | 6.34 | 0.60 |
| 2 | 6.11 | 5.67 | 0.44 |
| 3 | 6.31 | 5.79 | 0.52 |
| 4 | 6.35 | 5.66 | 0.69 |
| 5 | 6.21 | 5.74 | 0.47 |
| 6 | 6.17 | 5.71 | 0.46 |
| 7 | 6.17 | 5.69 | 0.48 |
| 8 | **_6.10_** | **_5.62_** | 0.48 |
| 9 | 6.18 | 5.63 | 0.55 |
| 10 | 6.24 | **_5.62_** | 0.62 |