@i-am-shastin 2025-02-18
Seems like I've missed the fact that Protobuf repo should be checked out before the cargo build/test. Sorry!

Btw I'm going to update the code anyway, implemented some kind of auto-generation for weapon skins data as well.